refactor: form logic
