+++ src/v2/Apps/SavedSearchAlerts/Routes/Overview/types.ts
@@ -1,5 +1,3 @@
-import { SavedSearchAleftFormValues } from "v2/Components/SavedSearchAlert/SavedSearchAlertModel"
-
 export interface EditAlertEntity {
   id: string
   name: string
@@ -7,9 +5,8 @@ export interface EditAlertEntity {
 }
 
 export interface EditFormProps {
-  initialValues: SavedSearchAleftFormValues
   editAlertEntity: EditAlertEntity
-  onSubmit: (values: SavedSearchAleftFormValues) => void
   onCloseClick: () => void
   onDeleteClick: () => void
+  onCompleted: () => void
 }
